05.01Update
Use Toon shader create the scene in Blender.
Q: Can not bake. There is "Shader to RGB" node in the shader
and It will probably have "Diffuse BDSF" connected to it,
Because most Toon shader will always use that and it's the reason why you cannot bake it normally because you can only bake in Cycle and Cycle doesn't support  "Shader to RGB".
A: Trying to solve the problem.